requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
python-dotenv==1.0.1
pydantic==2.12.5
starlette==0.36.3

pymongo==4.15.5
supabase==1.0.4
requests==2.32.5

numpy==1.23.5
pandas==1.5.3
scikit-learn==1.7.2
scipy==1.15.3
numba==0.62.1
librosa==0.10.1
soundfile==0.12.1
sounddevice==0.5.3

mediapipe==0.10.8
deepface==0.0.79
mtcnn==0.1.1
retina-face==0.0.17

torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
transformers==4.36.2
tokenizers==0.15.2
huggingface-hub==0.36.0

opencv-python-headless==4.7.0.72

wordninja==2.0.0
pyspellchecker==0.8.4
tqdm==4.64.1
regex==2025.11.3

beautifulsoup4==4.14.3
soupsieve==2.8
python-dateutil==2.9.0.post0
colorama==0.4.6
filelock==3.19.1
joblib==1.5.2
packaging==25.0
protobuf==3.20.3
typing_extensions==4.15.0
gdown==4.7.1
